requirements.txt
flask
pandas
scikit-learn
python-dotenv
groq
chartjs
numpy